+++ ReadingOutputFromFLYonPIC_v2/ChargeStateColors.py
@@ -13,25 +13,3 @@ import typeguard
 import typing
 from . import Config as cfg
 
-@typeguard.typechecked
-def getChargeStateColors(config : cfg.AtomicPopulationPlot.PlotConfig | cfg.ChargeStatePlot.PlotConfig, additionalIndices : list[int] = []):
-    """@return dictionary assigning one color to each charge state"""
-    colors = iter([config.colorMap(i) for i in range(config.numColorsInColorMap)])
-
-    ## assign all chargeStates a color
-    colorChargeStates = {}
-    for z in config.chargeStatesToPlot:
-        try:
-            colorChargeStates[z] = next(colors)
-        except StopIteration:
-            colors = iter([config.colorMap(i) for i in range(config.numColorsInColorMap)])
-            colorChargeStates[z] = next(colors)
-
-    for index in additionalIndices:
-        try:
-            colorChargeStates[index] = next(colors)
-        except StopIteration:
-            colors = iter([config.colorMap(i) for i in range(config.numColorsInColorMap)])
-            colorChargeStates[index] = next(colors)
-
-    return colorChargeStates
